Fix for Number positioning bug (Ticket #13 in the original repo).
I have fixed the positioning of the numbers, even though this bug is caused by FF10 not correctly implementing the CSS specs.
If a child is positioned absolutely and it's parent is positioned relatively, the child should be positioned relatively to it's parent... Apparently FF doesn't think this counts when the parent is a TD tag and decides to position relatively to the TABLE tag. Oh well.